# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Powered E-Commerce Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  buyer: "buyer",
  seller: "seller",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  role: RegisterBodyRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  buyer: "buyer",
  seller: "seller",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  isApproved: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type UpdateUserRoleBodyRole =
  (typeof UpdateUserRoleBodyRole)[keyof typeof UpdateUserRoleBodyRole];

export const UpdateUserRoleBodyRole = {
  buyer: "buyer",
  seller: "seller",
  admin: "admin",
} as const;

export interface UpdateUserRoleBody {
  role: UpdateUserRoleBodyRole;
  isApproved?: boolean;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
}

export type ProductStatus = (typeof ProductStatus)[keyof typeof ProductStatus];

export const ProductStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface ProductImage {
  id: number;
  productId: number;
  url: string;
  isPrimary: boolean;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  stock: number;
  /** @nullable */
  categoryId?: number | null;
  sellerId: number;
  status: ProductStatus;
  images?: ProductImage[];
  category?: Category;
  seller?: User;
  createdAt: string;
}

export interface ProductsPage {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface CreateProductBody {
  name: string;
  description: string;
  price: number;
  stock: number;
  /** @nullable */
  categoryId?: number | null;
}

export interface UpdateProductBody {
  name?: string;
  description?: string;
  price?: number;
  stock?: number;
  /** @nullable */
  categoryId?: number | null;
}

export interface AddProductImageBody {
  url: string;
  isPrimary?: boolean;
}

export type UpdateProductStatusBodyStatus =
  (typeof UpdateProductStatusBodyStatus)[keyof typeof UpdateProductStatusBodyStatus];

export const UpdateProductStatusBodyStatus = {
  approved: "approved",
  rejected: "rejected",
} as const;

export interface UpdateProductStatusBody {
  status: UpdateProductStatusBodyStatus;
}

export interface CartItem {
  productId: number;
  quantity: number;
  product?: Product;
  subtotal: number;
}

export interface Cart {
  items: CartItem[];
  total: number;
  itemCount: number;
}

export interface AddToCartBody {
  productId: number;
  quantity: number;
}

export interface UpdateCartItemBody {
  quantity: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface OrderItem {
  id: number;
  orderId: number;
  productId: number;
  quantity: number;
  unitPrice: number;
  product?: Product;
}

export interface Order {
  id: number;
  buyerId: number;
  status: OrderStatus;
  totalAmount: number;
  shippingAddress: string;
  items?: OrderItem[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateOrderBody {
  shippingAddress: string;
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  confirmed: "confirmed",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export type ReturnStatus = (typeof ReturnStatus)[keyof typeof ReturnStatus];

export const ReturnStatus = {
  pending: "pending",
  verifying: "verifying",
  approved: "approved",
  rejected: "rejected",
  manual_review: "manual_review",
} as const;

/**
 * @nullable
 */
export type ReturnAiVerdict =
  | (typeof ReturnAiVerdict)[keyof typeof ReturnAiVerdict]
  | null;

export const ReturnAiVerdict = {
  PASS: "PASS",
  FAIL: "FAIL",
  SUSPICIOUS: "SUSPICIOUS",
} as const;

export interface ReturnImage {
  id: number;
  returnId: number;
  url: string;
}

export type VerificationResultVerdict =
  (typeof VerificationResultVerdict)[keyof typeof VerificationResultVerdict];

export const VerificationResultVerdict = {
  PASS: "PASS",
  FAIL: "FAIL",
  SUSPICIOUS: "SUSPICIOUS",
} as const;

export type VerificationChecksFraudRisk =
  (typeof VerificationChecksFraudRisk)[keyof typeof VerificationChecksFraudRisk];

export const VerificationChecksFraudRisk = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface VerificationChecks {
  productIdentityMatch: boolean;
  packagingIntact: boolean;
  noVisibleDamage: boolean;
  correctProduct: boolean;
  fraudRisk: VerificationChecksFraudRisk;
}

export interface VerificationResult {
  /** @nullable */
  id?: number | null;
  returnId: number;
  verdict: VerificationResultVerdict;
  similarityScore: number;
  checks: VerificationChecks;
  /** @nullable */
  createdAt?: string | null;
}

export interface Return {
  id: number;
  orderId: number;
  orderItemId: number;
  buyerId: number;
  reason: string;
  status: ReturnStatus;
  /** @nullable */
  aiVerdict?: ReturnAiVerdict;
  /** @nullable */
  aiScore?: number | null;
  /** @nullable */
  adminNotes?: string | null;
  images?: ReturnImage[];
  verificationResult?: VerificationResult;
  createdAt: string;
  updatedAt: string;
}

export interface CreateReturnBody {
  orderId: number;
  orderItemId: number;
  reason: string;
}

export interface AddReturnImageBody {
  url: string;
}

export type UpdateReturnDecisionBodyStatus =
  (typeof UpdateReturnDecisionBodyStatus)[keyof typeof UpdateReturnDecisionBodyStatus];

export const UpdateReturnDecisionBodyStatus = {
  approved: "approved",
  rejected: "rejected",
} as const;

export interface UpdateReturnDecisionBody {
  status: UpdateReturnDecisionBodyStatus;
  adminNotes?: string;
}

export interface SellerStats {
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  pendingReturns: number;
  approvedReturns: number;
}

export interface AdminStats {
  totalUsers: number;
  totalSellers: number;
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  totalReturns: number;
  fraudAlertsCount: number;
  pendingApprovals: number;
}

export type ListProductsParams = {
  search?: string;
  categoryId?: number;
  minPrice?: number;
  maxPrice?: number;
  page?: number;
  limit?: number;
};